Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int a,b;
-    scanf("%d %d",&a,&b);
+    float a,b;
+    scanf("%f %f",&a,&b);
     if (a/b==b){
         printf("Yes");
     }else{
